Modify the interpolation of G(T) from Chianti
With this commit, I have modified the interpolation of the G(T) values
from the CHIANTI tables. It used the CGAL interpolation, but that was
very slow. Now, we directly look up the correct point, where explicitly
use the order of storage in the chiantitables.

The new interpolation looks up the index in the density vector, and does
a linear interpolation in the temperature direction.