Added the Projection rendering method
The Projection rendering method assumes that all points/voxels in the
datacube have (nearly) equal volume and just projects them on the
plane-of-sky. The method is very fast, but suffers from pixelation
issues. The projection method cannot correctly estimate the path length in the
voxels: a warning is given that the user should renormalise the output
to obtain the absolute intensity values.