Fixed boost detection if it was installed to user directory
Xiaozhou's installation failed because of the bad detection of boost. He compiled boost and installed it to a separate directory with make install. Now I just added $ac_boost_path/include to the include directories. Then it automatically picks up the headers and libs, when the install dir is specified with --with-boost.