Experimental version of FoMo for Thomson scattering
Implemented necessary functions to integrate emission of Thomson
scattering. To that end, I included a new function in fomo-thomson.cpp
to have emission that depends on LOS.
Probably, the computation of the emission needs to be done in a separate
routine, similar to fomo-CHIANTI.cpp. Now it requires a Goftcube to be
set, before it is integrated.