Updated FoMo-C to take into account the units of emission
The interpolation method now uses the units of the chiantitables as a
decider on what to output. If the unit string contains DN, then we are
talking about an imager, and everything is returned in arcsec and
DN/pixel/s. Otherwise, we are modelling a spectral line, and then the
spatial coordinates are using Mm, and the intensity has the appropriate
units as well.