Update the configure script for more modern installations.
The configure script did no longer successfully run for modern versions
of boost, autoconf and automake. Because of changes in boost, I also had
to change some examples making use of the directory_iterator in boost. I
switched to the std::filesystem one. Perhaps some bugs still exist in the
directory searching in the examples, because I have not tested it yet.